next period start estimation wording improvement
Currently what the page shows is this: `The next period starts in approximately about 18 hours (106 blocks)

"Approximately about" is of course incorrect, and it stems from the fact that date-range formatting library inserts the `about`. I think it's simpler just to avoid the approximately. 

Alternatively it could say "The next period is estimated to start in..." and then the "about" won't be a problem. My personal preference is for the above shorter phrasing.